get_post_by_user

+++ app/Http/Services/Impl/PostServiceImpl.php
@@ -5,16 +5,21 @@ namespace App\Http\Services\Impl;
 
 
 use App\Http\Repositories\PostRepository;
+use App\Http\Repositories\UserRepository;
 use App\Http\Services\PostService;
 use App\Models\Post;
 use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 
 class PostServiceImpl implements PostService
 {
     public $postRepository;
-    public function __construct(PostRepository $postRepository)
+    public $userRepository;
+
+    public function __construct(PostRepository $postRepository, UserRepository $userRepository)
     {
         $this->postRepository = $postRepository;
+        $this->userRepository = $userRepository;
     }
 
     public function getAll()
@@ -22,10 +27,17 @@ class PostServiceImpl implements PostService
         return $this->postRepository->getAll();
     }
 
-    public function findById($id)
+    public function destroy($id)
     {
         $post = $this->postRepository->findById($id);
-        return $this->postRepository->destroy($post);
+        $user = $post['user_id'];
+        if ( Auth::id() == $user){
+            return $this->postRepository->destroy($post);
+        }else{
+            $message = "Không được phép xóa bài viết";
+            return $message;
+        }
+
     }
 
     public function update($request,$id)
@@ -39,5 +51,18 @@ class PostServiceImpl implements PostService
     {
         $post = $this->postRepository->create($request);
         return $post;
+    }
+
+    public function findById($id)
+    {
+        $post = $this->postRepository->findById($id);
+        $user = $post['user_id'];
+        if ( Auth::id() == $user){
+            return $post;
+        }else{
+            $message = "Không được phép chỉnh sửa bài viết";
+            return $message;
+        }
+
     }
 }
